Convert an object to a period

Convert an object to a period

as_period(x, interval, firstdate, ...)

# S3 method for default
as_period(x, interval, firstdate, ...)

# S3 method for period
as_period(x, interval = 1L, firstdate = NULL, ...)

# S3 method for Date
as_period(x, interval = 1L, firstdate = NULL, ...)

# S3 method for POSIXt
as_period(x, interval = 1L, firstdate = NULL, ...)

# S3 method for character
as_period(x, interval = 1L, firstdate = NULL, ...)

# S3 method for factor
as_period(x, interval = 1L, firstdate = NULL, ...)

Arguments

x
An object to coerce to a period.

interval
An integer or character indicating the (fixed) size of the time interval used for computing the incidence; defaults to 1 day. More details can be found in the "Interval specification" below.

firstdate
The date to anchor the intervals from. If NULL (default) the earliest date in the vector will be used.

...
Not currently used.

Value

A period object except when interval values 1, 1L, "day" or "days" (both without prefix) are used. In this situation the returned object is of the standard "Date" class.

Interval specification

interval can be specified as either an integer value or a more standard specification such as "day", "week", "month", "quarter" or "year". The format in this situation is similar to that used by seq.Date() where these values can optionally be preceded by a (positive or negative) integer and a space, or followed by "s"

* (x) day(s)
* (x) weeks(s)
* (x) epiweeks(s)
* (x) isoweeks(s)
* (x) months(s)
* (x) quarter(s)
* (x) years(s)

Note that for the values "month", "quarter" and "year", intervals are always made to start at the beginning of the calendar equivalent. If the input is an integer value the input is treated as if it was specified in days (i.e. 2 and 2 days) produce the same output.

The only interval values that do not produce these grouped classes are 1, 1L, "day" or "days" (both without prefix) are used. In this situation the returned object is of the standard "Date" class.

Note

Internally period objects are represented by the date at the beginning of the period and then stored as the number of days (starting at 0) since the Unix Epoch (1970-01-01).

Examples

as_period(Sys.Date(), interval = "28 days")

#> <period> interval = 28 days
#> [1] "2021-03-16 to 2021-04-12"

as_period(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"), interval = 10)

#> <period> interval = 10 days
#> [1] "2019-03-04 to 2019-03-13"

as_period("2019-05-03", interval = "2 weeks")

#> <period> interval = 2 weeks
#> [1] "2019-04-29 to 2019-05-12"