<grates_month> scale

ggplot2 scale for <grates_month> vector.

Usage

scale_x_grates_month(
  n.breaks = 6,
  format = "%Y-%m-%d",
  bounds_format = "%Y-%b",
  sep = "to",
  n,
  origin
)

Arguments

n.breaks

Approximate number of breaks calculated using scales::breaks_pretty (default 6).

format

Format to use if "Date" scales are required. If NULL then labels are centralised and of the form "lower category bound to upper category bound". If not NULL then the value is used by format.Date() and can be any input acceptable by that function (defaults to "%Y-%m-%d).

bounds_format

Format to use for grouped date labels. Only used if format is NULL.

sep

Separator to use for grouped date labels.

n

Number of months used for the original grouping.

origin

Month since the Unix epoch used in the original grouping.

Value

A scale for use with ggplot2.