Construct a grates_yearweek object

yearweek() is a constructor for a <grates_yearweek> object.

yearweek(x = integer(), firstday = 1L)

is_yearweek(x)

Arguments

x

Integer vector representing the number of weeks.

firstday

An integer representing the day the week starts on from 1 (Monday) to 7 (Sunday).

<grates_yearweek> objects are stored as the number of weeks (startint at 0) from the date of the specified firstday nearest the Unix Epoch (1970-01-01). That is, the number of seven day periods from:

- 1969-12-29 for `firstday` equal to 1 (Monday)
- 1969-12-30 for `firstday` equal to 2 Tuesday
- 1969-12-31 for `firstday` equal to 3 Wednesday
- 1970-01-01 for `firstday` equal to 4 Thursday
- 1970-01-02 for `firstday` equal to 5 Friday
- 1970-01-03 for `firstday` equal to 6 Saturday
- 1970-01-04 for `firstday` equal to 7 Sunday

Examples

yearweek(0:5)
#> <grates_yearweek[6]>
#> [1] 1970-W01 1970-W02 1970-W03 1970-W04 1970-W05 1970-W06